backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
plaid-python==20.0.0
anthropic==0.34.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0
pandas==2.2.2
reportlab==4.2.2
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
